py175/lesson3/users_and_interests/pyproject.toml
[project]
name = "users-and-interests"
version = "0.1.0"
description = ""
authors = [
    {name = "flash-ost",email = "dotoner555@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
